Added content on home page. It was very painful.
The home page is almost working, except for the configurations section,
which sucks since presenting customers with pre-built configurations is
the primary purpose of the project.

I can't get the @/type/Configurations module to work yet and it causes a
lot of problems because of that, which is very weird since it has like 5
lines and no mistake was spotted.

Anyway, at least there's a home page with some content.

Next stop: The "view more" buttons.